+++ RecordTimes/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="hgsportfolio.example.RecordTime.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/activity_fragment_container"
