+++ RecordTimes/app/src/main/res/layout/fragment_modal_update_delete.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:layout_gravity="center"
     android:background="#FDFCFC"
-    tools:context=".UpdateDeleteModalFragment">
+    tools:context="hgsportfolio.example.RecordTime.UpdateDeleteModalFragment">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/closeUpdateModal"
